INKO-200 Add request to verify client Powered by JAXB

+++ fineract-provider/src/main/java/org/apache/fineract/portfolio/loanaccount/service/TransUnionCrbClientVerificationReadPlatformService.java
@@ -22,5 +22,5 @@ import org.apache.fineract.portfolio.loanaccount.data.TransUnionRwandaClientVeri
 
 public interface TransUnionCrbClientVerificationReadPlatformService {
 
-    TransUnionRwandaClientVerificationData retrieveClientToBeVerifiedToTransUnion(Integer clientId);
+    TransUnionRwandaClientVerificationData retrieveClientToBeVerifiedToTransUnion(Long clientId);
 }
